refac wmap into one file
